AUDIT-47: Replaced the custom logout listener with in build user session listener pattern
